+++ agate/data_types/__init__.py
@@ -1,18 +1,18 @@
 #!/usr/bin/env python
 
 """
-This module contains the :class:`.DataType` class and its subclasses. These
-types define how data should be converted during the creation of a
+Data types define how data should be imported during the creation of a
 :class:`.Table`.
 
-A :class:`TypeTester` class is also included which be used to infer data
-types from column data.
+If column types are not explicitly specified when a :class:`.Table` is created,
+agate will attempt to guess them. The :class:`.TypeTester` class can be used to
+control how types are guessed.
 """
 
 from copy import copy
 
 from agate.data_types.base import DEFAULT_NULL_VALUES, DataType  # noqa
-from agate.data_types.boolean import Boolean
+from agate.data_types.boolean import DEFAULT_TRUE_VALUES, DEFAULT_FALSE_VALUES, Boolean
 from agate.data_types.date import Date
 from agate.data_types.date_time import DateTime
 from agate.data_types.number import Number