+++ agate/data_types/boolean.py
@@ -21,13 +21,15 @@ class Boolean(DataType):
     """
     Data type representing boolean values.
 
-    Note: numerical `1` and `0` are considered valid boolean values, but no
-    other numbers are.
+    Note that by default numerical `1` and `0` are considered valid boolean
+    values, but no other numbers are.
 
     :param true_values: A sequence of values which should be cast to
-        :code:`True` when encountered with this type.
+        :code:`True` when encountered with this type. Defaults to
+        :data:`.DEFAULT_TRUE_VALUES`.
     :param false_values: A sequence of values which should be cast to
-        :code:`False` when encountered with this type.
+        :code:`False` when encountered with this type. Defaults to
+        :data:`.DEFAULT_FALSE_VALUES`.
     """
     def __init__(self, true_values=DEFAULT_TRUE_VALUES, false_values=DEFAULT_FALSE_VALUES, null_values=DEFAULT_NULL_VALUES):
         super(Boolean, self).__init__(null_values=null_values)